requirements.txt
Django==4.0.1
celery==4.4.7
redis==4.1.0
coverage==5.0.4
djangorestframework==3.13.1
django-filter==21.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
django-environ==0.8.1
dj-database-url==0.5.0
psycopg2-binary==2.9.3